prov/efa: Handle CUDA buffers in the send path
This commits adds a helper function rxr_copy_to_tx()
which can tell if a registration is of type
FI_HMEM_CUDA and uses the CUDA IOV copy util in the send path to copy
out of the device buffer as needed.

It also rename the following function:

    rxr_pkt_send_data_mr_cache()
    rxr_prepare_mr_send()

to
    rxr_pkt_send_data_desc()
    rxr_prepare_desc_send()

To reflect the fact that these functions were also used when app
provides mem_desc.

Signed-off-by: Raghu Raja <craghun@amazon.com>
Signed-off-by: Wei Zhang <wzam@amazon.com>