prov/shm: fix potential data ordering issue in atomic fetch path
Remove an optimization that allows for some atomic
fetch operations to remotely fetch the data from the
peer using CMA before the peer actually processes the
incoming request. To be completely correct, this would
have to have a message ordering check around it to make
sure the application does not require atomic write after
read ordering. However, it is a rare use case that makes
the code more complicated so this is getting removed
rather than getting fixed to check the ordering.

This patch removes the optimization, making all atomic
fetch operations use the SMR_RMA_REQ path and also cleans
up and simplifies the code to make it more readable.

Note that applications using previous versions may be
broken if data ordering for atomic fetches is needed.

Signed-off-by: aingerson <alexia.ingerson@intel.com>