prov/tcpx: Remove incorrect events reported to EQ
When a connection is shut-down, the EQ may receive
incorrect events for failed RX entries. This may
cause some bad bugs in the user application because
a single call to fi_shutdown() may generate multiple
EQ events for the same endpoint.

The patch removes the call to tcpx_report_error(), which
incorrectly causes invalid events to be reported in
the EQ. Most significantly, tcpx_report_error() doesn't
protect from reporting multiple FI_SHUTDOWN events
for the same endpoint.

Removing this function is safe because the call to
ep->get_rx_entry[op](ep) (which maps to the functions
tcpx_get_rx_entry_op_*) already does the following
in case the RX entry fails:
1. It generates a CQ event.
2. It cleans the RX entry and reset ep->cur_rx_entry
   to NULL. The new assert() enforces this assumption.

Signed-off-by: Sylvain Didelot <sdidelot@ddn.com>