core/mr_cache: Remove region merging support
Merging requires that we atomically find all entries that overlap
with a given range, remove them from the cache, and insert the
new region in its place.  However, we can't safely hold a lock on
the cache when allocating or freeing memory, since doing so could
invoke the monitor callback handling.

Since we need to be able allocate and register new regions without
holding a lock on the cache, support for merging entries becomes
impractical.  Although we can rely on the memory backing the new
region to remain, memory backing the region that we're mering with
may be invalidated at any time.

To prepare the code to build new regions without holding the cache
lock, we remove support for region merging.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>