prov/rxm: Report CQ errors for failed RMA transfers
RxM currently only reports errors for a subset of data transfers that
fail.  For others (such as RMA), it reports an incomplete, generic
error.  Improve error reporting.  Note that some transfers types are
still not being reported.  A full update will be needed as a follow on
patch set.

As part of this change, we restructure successful completion reporting,
so that the switch case statements align with the order of the transfer
states.  This helps ensure that all states are covered.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>