configure: Support configuring for ROCr
ROCr provides an HSA interface which includes AMD GPU devices. The ROCr
HSA interface enables libfabric to integrate with AMD GPU devices. This
includes moving data between host and device buffers, and a device
registration cached.

Note that if a libfabric client is using ROCm HIP configured for an HCC
platform, the ROCr interfaces are compatible.

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>