prov/efa: call rxr_pkt_insert_addr() if REQ packet contain raw address
Currently rxr_pkt_insert_addr() is called only when the source address
is FI_ADDR_UNAVAIL, this behavior will cause problem when the packet
is from a newly created EP with the same GID and QP of a previous one.

This patch address the issue by calling rxr_pkt_insert_addr() as long
as incoming REQ packet contains raw address in the header.

Signed-off-by: Wei Zhang <wzam@amazon.com>