prov/verbs: Add check against domain name
Commit 77b79ae removed the domain name check from the util
code in order to handle dynamically named domains (e.g. tcp
and udp providers).  However, verbs needs this check when
multiple devices are present, or we can match with the
incorrect device.  (Verbs forms a list of all valid domains
during initialization).  Add back in the check but directly
in the verbs code when it's searching its saved fi_info domain
list.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>