core: Define new wait object type FI_WAIT_POLLFD
When dealing with socket based providers (tcp, udp), wait calls must
check the state of multiple file descriptors.  This is currently
only possible when epoll is available, and all fd's are added to an
epoll fd.  That allows the provider to support FI_WAIT_FD.

However, epoll may not be available, or it may be preferable to not
use epoll for performance reasons.  Internally, libfabric already
handles this via an epoll abstraction that implements epoll semantics
over poll.  However, that implementation cannot be exposed to
users through the API.

Introduce a more generic wait object, POLLFD, that allows an OFI wait
object to be backed by an array of fd's.  That is, FI_WAIT_POLLFD
becomes a superset of FI_WAIT_FD.  The native wait object in this
case can be reported as an array of struct pollfd.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>