prov/util: Add wait set support for FI_WAIT_POLLFD
Implement a wait set to handle FI_WAIT_POLLFD.  This is a slight
derivation from the wait set to support a single fd.  Unfortunately,
because the function signatures differ between epoll and struct
pollfds, we need to invoke the correct ofi_epoll_xxx or ofi_pollfds_xxx
call based on the wait set that was requested.  And we don't want
to copy-paste-modify the entire_wait_fd functionality, so we insert
checks against the wait object to call the correct function.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>