util/wait: Cleanup add/del fid helper functions
The yield based wait set allows associating a fid with a wait
set.  That same functionality will be needed for fd based wait
sets.

This updates those helpers to take the fid as input (rather than
a void * pointer), renames them, and prepares the calls so
other wait sets can use the functions.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>