core/wait: Add new fi_control() command to get the wait object type
Provide a mechanism for the user to discover the type of underlying
wait object it is dealing with.  This will allow the underlying
provider to select the best option for its use, while allowing the
upper caller to adjust based on the selected value.

This will be used by util wait code to obtain the correct wait object
of fid's associated with the wait set.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>