core: Define generic provider version mechanism
Provider versions are not getting updated as new libfabric versions are
being released.  To fix this moving forward, the following changes are
being proposed.

A new FI_REVISION_VERSION value is added to fabric.h.  This value will
align with the last number in the version release.  That is, for version
vX.Y.Z, it will correspond to the Z value.

By default, provider versions will be set to a value that corresponds with
Y.(Z+10) as their version.  The provider major value will be the minor value for
libfabric, and the provider minor value will be the stable release
revision, offset by 10.  E.g. 10.0, 10.10, 10.20, etc.

The offset allows for custom builds of libfabric to set a provider version
different from that used by the mainline releases.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>