prov/efa: Fix addr_format handling for RDM endpoints
The emulation layer for RDM endpoints in the EFA provider was
overwriting the addr_format field of the caller's hints without
verifying the contents.  This would result in endpoints being
returned with address formats other than the one requested (for
example, requesting FI_ADDR_STR would result in an endpoint with
FI_ADDR_EFA address format).

This patch removes that munging.  The core efa provider already
has a check that the addr_format is either EFA or UNSPEC, so the
end result of this patch is that efa endpoints will not be
returned if the addr_format in hints is not EFA or UNSPEC.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>