prov/efa: handle the case an EP want to use RDMA read on itself
Currently, an EP will use RDMA read on a peer only if both the EP
and the peer support RDMA read. The peer's information was received
via a handshake packet.

In the case an EP want to use RDMA read on itself, it does not
need to wait for a handshake packet. If should use RDMA as long
as it supports RDMA.

This patch introduced a flag is_self to struct rxr_peer, which
was set at efa_av_insert_addr(), and was used to determine whether
RDMA read should be used.

Signed-off-by: Wei Zhang <wzam@amazon.com>