prov/shm: allow building of shm provider on older kernels
When building on an older kernel without CMA for use on
systems with CMA, shm will not get built and will be disabled.

To bypass this for x86 systems, skip the configure check and
provide a wrappers for the CMA calls which call the syscall
directly.

The newly added runtime checks for CMA calls and the alternate
large message protocol for use without CMA will allow the shm
provider to run with or without CMA.

Signed-off-by: aingerson <alexia.ingerson@intel.com>