prov/efa: fix a bug in rxr_pkt_send_data()
A bug was introduced when we introduced the function
     rxr_copy_from_tx()
The return value of that function was not checked
and there is a wrong assertion in the code.

This function was only used when MR cache was turned off,
which is why the code has been passing the tests.

This patch addressed the issue.

Signed-off-by: Wei Zhang <wzam@amazon.com>