prov/psm2: Remove obsolete comments
In the past, FI_REMOTE_CQ_DATA was a capability bit and the provider
used it to determine if a provider instance with TAG64 layout should
be enabled.

This is no longer true and the check was removed by commit 8b9ae5ba8dc.
However, some comments were left behind and they make no sense under
current context.

Remove those obsolete comments and add back the scenarios that were
previously covered by the caps bit for enabling the TAG64 layout.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>