prov/shm: reorganize progress_resp_entry
Previously, the freeing of the tx_buf was outside of
the switch and all non-inject cases were ending in
a goto out to skip the tx_buf freeing.

Make all cases break instead of goto out and only
free the tx_buf if it was used.

This makes the function clearer, more intuitive, and
easier to add other cases where freeing is necessary

Signed-off-by: aingerson <alexia.ingerson@intel.com>