mrcache: Introduce CUDA buffer invalidation checks
This commit extendes the cached entry validation in
ofi_mr_cache_search/find() to include a CUDA buffer ID check, which
helps catch deallocations by the application. In order to do this, the
original registration also includes a query to CUDA to get the base ID
for an allocation, when creating an entry in util_mr_cache_create().

Proof-of-concept change to help discussion in:
https://github.com/ofiwg/libfabric/issues/5789

Signed-off-by: Raghu Raja <craghun@amazon.com>