prov/shm: minimize locking in progress_resp_entry
The peer smr region only needs to be locked when freeing
the region's resources. If it's locked, the peer cannot
make any progress on any transfers either.

Moving this to the end minimizes locked code and allows
for SAR transfers to make use of their 2 buffers
simultaneously.

A few extra checks are needed to make sure we are not
processing a transfer twice if it was not able to
complete last time due to the try_acquire.

Signed-off-by: aingerson <alexia.ingerson@intel.com>