prov/efa: fix memory leak in rxr_rma_post_shm_write()
rxr_rma_post_shm_write() allocate pkt_entry and use it
as context to call fi_writemsg() on shm_ep. In the
case that fi_writemsg() failed, it did not release the
pkt_entry, causing a memory leak.

This patch address the issue by release pkt_entry in
case fi_writemsg() return an error.

It also do a little refactoring: the function name

       rxr_rma_port_rtw()

is not correct, because it not only post rtw, but also
call rxr_rma_post_shm_write(). This patch rename it to

       rxr_rma_post_write().

Signed-off-by: Wei Zhang <wzam@amazon.com>