util/util_attr: fix ofi_cap_mr_mode to not drop FI_MR_HMEM
FI_MR_HMEM is similar to FI_MR_LOCAL, in that it specifies that device buffers
must be registered. However, it is valid for a provider to require FI_MR_HMEM
support but not FI_MR_LOCAL.

Change ofi_cap_mr_mode to not drop FI_MR_HMEM so this mr mode bit is still set
in this case.

Signed-off-by: Robert Wespetal <wesper@amazon.com>