prov/tcp: Do not signal wait object after adding fd
This fixes an issue with fi_cq_sread failing with ECANCELED because
of an incorrect signaling of the wait fd.  When a new fd is added to
a wait set, the wait set will wake up any waiting thread internally,
if needed.  Otherwise, signaling the wait set should be reserved for
waking up a waiting thread because a corresponding event or completion
has been written.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>