prov/util: Fix return code in ofi_cq_sreadfrom on signal
If a thread is signaled while blocked in cq sread, the return value
should be EAGAIN, instead of ECANCELED.  This aligns with the man
page, but also fixes a race where a new entry can be added to the CQ,
and the cq signaled to wake up any thread.  A thread reading the CQ
may read the completion prior to blocking.  The next time the thread
tries to read the CQ, it will find the CQ empty, but the signal set.
This results in the thread returning with an ECANCELED error.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>