prov/efa: Fix error handling when av open fails
Before this fix, we go to efa_av_close when
shm's av_open fails and dgram's av->conn_table
calloc fails. This is not correct, because efa's
av has not been opened successfully.

In this patch instead, we close the resources which
are opened/initialized till the failure point
(shm_av, ofi_util_av, efa_av).

Signed-off-by: Jie Zhang <zhngaj@amazon.com>