prov/shm: fix locking when processing response from self
When sending a message to itself and processing the response
from a peer, the peer_smr is the same as the local smr so it
is impossible to successfully acquire the second lock since
it has already been acquired.

Add a check when processing the responses to make sure we
only acquire other locks, and not our own.

This also requires skipping mapping of peers within the same
process, both for avoiding unnecessary mapping and for saving
the local region pointer, allowing comparison of the pointers
instead of region names.

Signed-off-by: aingerson <alexia.ingerson@intel.com>