prov/psm2: Improve source address translation for scalable endpoints
The 'source' address information of a message received from the PSM2
layer contains only the PSM2 endpoint address. The information itself
is insufficient to tell whether the source is a scalable endpoint. If
source has already been inserted into the address vector, the provider
can retrieve the right fi_addr by looking into the addresses of each
SEP context. However, if the source has not been inserted into the
address vector, the generated FI_EADDRNOTAVAIL event have to assume the
address is a regular endpoint. This works fine if the address is then
inserted into the address vector and used as-is for later communication,
but can cause error if the application takes advantage of prior
knowledge and tries to use the resulting fi_addr as scalable endpoint
address.

This is improved by overloading the remote CQ data with sep_id. When
remote CQ data is unused, the space can be used to pass the sep_id of
the sender to the receiver. The receiver can now construct correct
source address for scalable endpoints. It falls back to the old way if
the remote CQ data is in use.

This feature is used by the RPC framework Mercury, which conveniently
does not need remote CQ data.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>