prov/verbs: vrb_get_rdma_rai always requires rai parameter
Comments in vrb_get_rdma_rai suggest that the caller does not have to
provider the rai parameter, which returns a list of rdma_addrinfo
structures.  However, if rai is not given, then there is a memory
leak of those structures.  Since all callers (only 2) provide this
parameter, update the comments and require that rai be passed in.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>