prov/rxm: Fix rxm_buffer_size param checking
If rxm_buffer_size and rxm_sar_limit size are set significantly low such
that most small message operations are issued using the rendezvous
protocol, the target receive buffer may be too small to hold the
incoming rendezvous control message. This results in the core provider
raising a CQ error which causes RxM to crash on an assert. To fix this
issue, always ensure that the rxm_buffer size is at least the size of a
rendezvous control message.

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>