prov/sockets: check connection's valid in sock_ep_lookup_conn
If peer with restart, it possibly being disconnect but not detected yet
when communicate with the peer. That leads to lookup invalid connection
and further cause failure in sock_pe_progress_tx_entry().

For example two nodes A and B, A send RPC request to B, and B will send
reply to A. But A only insert B's address to AV at first time and reuse
it later. After several RPC round trip, B restart executing with the same
IP+port. Then A cannot send RPC request to B any more because it always use
the cached fi_addr and cached connection information.

Another example - there are 3 nodes -- client_A and server_A, server_B, with
the usage steps as below -
1) client_A sends RPC+bulk_handle to server_A,
2) server_A finds that both server_A and server_B need to involve, so
   forwards the RPC+BULK to server_B
3) both server_A and server_B do bulk transferring from client_A, and
   server_B reports completions to server_A
4) server_A replies client_A the RPC's completion
In above step 3), to make server_B be able to start RMA transferring
from client_A, server_B need to insert client_A's address to AV to get
back usable fi_addr_t.
But server_B cannot insert it to AV every time because it will consume
one AV slot (sock_check_table_in), so in server_B at libfaric's upper
layer there is address cache that for same client only needs to insert
to AV at the first time, and lookup/reuse it later.
Then will with a problem that if client_A restarts executing and with
same IP+Port, the same executing of above 4 steps will cause the RMA
transferring's failure, because the cached fi_addr and its connection
information within socket provider is broken and invalid.

This patch fixes it by verifing that the existing connection is still
usable, and that the peer didn't restart, when lookup connection.

Signed-off-by: Xuezhao Liu <xuezhao.liu@intel.com>