prov/efa: fix logic of writing completion entry
This patch fix the logic of writing completion entry in
rxr_cq_write_tx_completion().

Currently it calls ofi_need_completion() to check if
completion is needed, but ofi_need_completion() is not
intended used here because it only check cq_flags and
op_flags.

This patch introduce

    rxr_cq_need_tx_completion() and

which check util_ep->tx/rx_msg_flags to see if
FI_SELECTIVE_COMPLETION was used. If so, they
check tx/rx_entry->fi_flags to see if completion should
be wrote.

Signed-off-by: Wei Zhang <wzam@amazon.com>