prov/efa: fix a bug in rxr_pkt_handle_recv_completion()
Currently rxr_pkt_handle_recv_completion() will call
    rxr_pkt_proc_common_req_hdr(),
if the received packet is an REQ packet.

This is wrong because the received REQ packet might be
from a different version of EFA provider, and the provider
cannot process the header.

This patch handle this case by printing a warning and
calling rxr_cq_handle_cq_error(), the same way we handle
an invalid control packet.

Signed-off-by: Wei Zhang <wzam@amazon.com>