prov/efa: reuse protection domain in same process address space
We currently are limited in the number of protection domains for EFA,
which causes issues with newer versions of Open MPI that open a
Libfabric domain in both the OFI MTL and BTL for a single rank.

Although we plan to increase this limit in the future, we have to add a
workaround to reuse a protection domain if possible in the meantime. Add
a list of protection domains and reuse it during domain open.

Signed-off-by: Robert Wespetal <wesper@amazon.com>