prov/efa: allow end point to choose shm usage
Currently, whether to use shm was controlled by a global variable
    rxr_env.enable_shm_transfer
, which means all end points behavior the same on shm usage.
However, some applications want to be able to choose shm usage when
it is enabled.

This patch allow application to do that by setting hints->caps
with FI_REMOTE_COMM and without FI_LOCAL_COMM off.

The following change has been made to the code to acheive that:

1. a flag use_shm has been added to struct rxr_ep, which will be
   set in rxr_endpoint()

2. only set peer->is_local flag when ep->use_shm is true.

3. In data transmission code path, change the condtion check:

        if (rxr_env.enable_shm_transfer && peer->is_local)

   to
        if (peer->is_local) with assert(ep->use_shm)

Signed-off-by: Wei Zhang <wzam@amazon.com>