core/mr_cache: Store HMEM iface type in MR info
The MR HMEM iface is now stored with the MR cache entry. This allows the
MR cache to select the correct memory monitor which is required for
multiple memory monitors per MR cache, and providers can identify the
HMEM iface type of a memory region being added to the cache.

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>