core/mr_cache: Define valid memory monitor ops
The valid memory monitor operation is used by the MR cache to query into
the monitor to see if the monitor's view of the virtual address is still
valid for DMA. For example, a memory monitor could store an identifier
between the pages and a virtual address. If the pages change for the
virtual address, the memory monitor's view of the virtual address is no
longer valid and the MR cache must unsubscribe and resubscribe to the
virtual address.

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>