cuda_memory_monitor: Initial CUDA memory monitor
The CUDA memory monitor cache assigns a buffer ID to each CUDA device
buffer. This buffer ID is unique based on device virtual address and the
underlying device pages. If the device pages behind a virtual address
change, a new buffer ID is generated. If the cached buffer ID is
different for the new buffer ID, the cached registration needs to be
evicted and remapped.

Note that the CUDA memory monitor requires providers to pin the device
buffer before an entry is inserted into the cache.

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>