prov/rxm: Replace handle_seg_data callback with direct call
A callback handler is defined for handle_seg_data_rx as a txrx_ops.
But this is only ever set to a single function.  Remove the handler
and just call the function directly.

Rename the function and make it static, as it no longer needs to be
exposed outside of the .c file.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>