prov/rxm: Replace handle_rndv callback with direct call
A callback handler is defined for handle_rndv_rx as part of the
handle_txrx_ops, but is only set to a single function.  Call that
function directly and remove the callback.

Rename the function and make it static as it no longer needs to be
exposed outside of the .c file.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>