prov/tcp: Simplify getting the next posted receive
tcpx_srx_next_xfer_entry obtains the next rx entry posted to
the shared rx.  At the same time, it updates the iov based
on the received message size.  Move the updating of the iov
outside of the function, so that it performs the single task
of getting the next entry.  Rename the function to clarify
what it does.

Provide a matching function to getting the rx entry when a
non-shared rx queue is used instead.

Move updating the iov, which is common to both cases, to a
a single location, as part of the rx_setup.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>