prov/rxm: Add verbs and tcp specific fi_info structs
To support verbs and/or tcp specific settings, associate specific
fi_info structs with those providers.  As these are the only in-tree
providers that rxm uses, position those first in the base rxm fi_info
list.  Following those are the generic fi_info structs for any
provider.  Yes, this will result in duplicate responses for tcp and
verbs, but other than the extra work, it is harmless.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>