prov/tcpx: Fix wrong error handling in fi_passive_ep()
Lifbabric doesn't report the correct error code to the
caller if fi_passive_ep() fails.

For example, in the case where the port is already bound
to another process, fi_passive_ep() would return -1 to
the application. It's because the value returned by bind()
is forwarded to the caller instead of the errno value.
(on error, bind() returns -1, and errno is set appropriately).

The patch also fixes an issue where fi_dupinfo() may fail
but fi_passive_ep() would return FI_SUCCESS.

Change-Id: Ie0244242c2d7020a3b0088b80d72be89025cceca
Signed-off-by: Sylvain Didelot <sdidelot@ddn.com>