src/hmem, prov/efa: pass device identifier into copy calls
Some device APIs may need or want to know which device they are
copying to or from (L0). Pass this in from the provider through the
stack to the copy call.

Note: this also moves the hmem_ops copy parameter iface to before
the iov (as well as the new device parameter) for clarity.

Change efa copy calls to match (and fix some indentation)

Signed-off-by: aingerson <alexia.ingerson@intel.com>