core/mr_cache: Define mem monitor start/stop ops
Two new memory monitor operations are defined: start and stop. The start
operation allows a memory monitor to allocate and setup internal
resources for the memory monitor to function. The stop operation allows
a memory monitor to tear down these resources.

The start operation must be called the first time an MR cache utilizes a
memory monitor. The stop operation must be called when no MR caches are
utilizing a memory monitor.

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>