prov/shm: redefine shm addrlen to not use NAME_MAX
NAME_MAX (256) is defined as not including the terminating
null character in a string.

While the shm provider does use this for strings, its
max length and addrlen refer to data size not string length
and must include the terminating character. Therefore,
NAME_MAX is not fully comparable to how shm uses it.

To fix this, define SMR_NAME_MAX as 256 and use this instead
to avoid confusion with the terms.

Signed-off-by: aingerson <alexia.ingerson@intel.com>