prov/efa: Reset tx_entry iov on Data Send Failure
This patch resets the tx_entry->iov_index and
tx_entry->iov_offset
in rxr_pkt_send_data_desc if rxr_pkt_entry_sendv fails.

We do this because the while loop that constructs
the iov is going to update the tx_entry->iov_index
and tx_entry->iov_offset.

If the send fails, we need to reset them.
Otherwise when we resend the packet,
loop conditions will not be satisfied and
thus iovs cannot be constructed.

Signed-off-by: Ao Li <aolia@amazon.com>