Fabtests: Replace python in runfabtests.sh with python3 or python2
In RHEL8, python is no longer linked to python 2. Instead, it is
preferred that the user specifies python2 or python3 (rather than
symlinking or setting alternatives). Since our scripts look to be
compatible with either version, we check for both and use python3 by
default.

This also exits the runfabtests script if python is not detected rather
than proceeding on (I have seen some unusual bugs when continuing without
python).

Signed-off-by: William Zhang <wilzhang@amazon.com>