core: Ensure that access to atomic handlers are in range
Unlike other 'last' values, OFI_SWAP_OP_LAST is actually a count, not the
last op + 1. Align the definition with other OP_LAST defines.

Add new OP_CNT defines to use where a count is actually needed.

Provide a general cleanup of how atomic constants and macros are used.
This ensures that all providers are consistent with various checks and
their use of atomics.  This fixes klocworks complaints that array
values may be out of range when accessing the atomic handlers.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>