prov/rxd: show correct msg_ordering
All packets are processed in order so all ordering is
supported. The following are exceptions:
FI_ORDER_RMA_WAR, FI_ORDER_SAR, FI_ORDER_WAR

This is because read requests trigger the target sending
data back so subsequent sends and writes might be out of order.
This does not apply to FI_ORDER_ATOMIC_WAR because the atomic
fetch data is buffered into another packet during the atomic
operation.

Signed-off-by: aingerson <alexia.ingerson@intel.com>